+++ features/support/app_pages/iva/firefighter_doc.rb
@@ -6,21 +6,33 @@ class DocFirefighter < IDmeBase
   include Capybara::DSL
   include ErrorMessages
 
-  def verify(populate = true)
+  def verify(populate: true, type: "none")
     find("[data-option=#{container_attribute}]").find(".verification-header").click
     populate_first_state(data_for(:experian_user).fetch("state"))
     choose("proboard_status_uncertified")
 
     if populate
-      populate_fields(data_for(:experian_user))
-      click_verify_button
-      attach_doc(2)
+
+      unique_data = data_for(:experian_user) #info for unique user
+      denied_data = data_for(:fail_experian) #info for denied user
+
+      case type
+      when "unique"
+        populate_fields(data: unique_data)
+      when "denied"
+        populate_fields(data: denied_data)
+      end
     end
 
     click_verify_button
+
+    if type == "unique"
+      attach_doc(2)
+      click_verify_button
+    end
   end
 
-  def populate_fields(data)
+  def populate_fields(data:)
     %w(first_name last_name social social_confirm street city).each do |field|
       fill_in field, with: data.fetch(field)
     end